Cleanup pong_tutorial_03 example
Move source code out of `src`.
Resolve path same way as other pong examples.
Add missing code snippet for adding system to ApplicationBuilder.
