+++ Cargo.toml
@@ -105,6 +105,10 @@ name = "pong_tutorial_02"
 path = "examples/pong_tutorial_02/main.rs"
 
 [[example]]
+name = "pong_tutorial_03"
+path = "examples/pong_tutorial_03/main.rs"
+
+[[example]]
 name = "appendix_a"
 path = "examples/appendix_a/main.rs"
