+++ examples/pong_tutorial_03/resources/bindings_config.ron
@@ -1,7 +1,7 @@
 (
   axes: {
     "left_paddle": (pos: Key(W), neg: Key(S)),
-    "right_paddle": (pos: Key(Up), neg: Key(Down))
+    "right_paddle": (pos: Key(Up), neg: Key(Down)),
   },
   actions: {},
 )
